feat: minimap in rogue mode

+++ src/SpriteSheet.cpp
@@ -21,7 +21,7 @@ namespace BloodSword::SpriteSheet
         {
             SDL_Rect rect;
 
-            auto height = int(std::nearbyint(Asset::Locations.size() / width + 0.5));
+            auto height = int(std::nearbyint((double)Asset::Locations.size() / (double)width + 0.5));
 
             rect.w = width * BloodSword::TileSize;
